+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Nix
   - Reduce Emanote's Nix runtime closure size
+- UI
+  - Add source map for Stork [\#391](https://github.com/EmaApps/emanote/pull/391)
 
 ## 1.0.0.0 (2022-12-04)