+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # Revision history for emanote
 
-## Unreleased
+## 1.0.2.0 (Unreleased)
 
 - Nix
   - Reduce Emanote's Nix runtime closure size
 - UI
   - Add source map for Stork [\#391](https://github.com/EmaApps/emanote/pull/391)
-- New features
+- Misc
   - Ignore toplevel `flake.{nix,lock}` by default.
+  - Remove deprecated `_emanote-bin/compile-css` script
 
 ## 1.0.0.0 (2022-12-04)